backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pynacl==1.5.0
base58==2.1.1
httpx==0.27.2
solders==0.21.0
python-dotenv==1.0.1
